NO-ISSUE Reorganize misc_spec.rb
